// app/backend/src/Interfaces/leaderBoard/ILeaderBoard.ts
export interface ILeaderBoard {
  name: string, // time
  totalPoints: number, // P
  totalGames: number, // J
  totalVictories: number, // V
  totalDraws: number, // E
  totalLosses: number, // D
  goalsFavor: number, // GP
  goalsOwn: number, // GC
  goalsBalance: number, // SG
  efficiency: string, // %
}